requirementswin.txt
absl-py==2.0.0
appdirs==1.4.4
asgiref==3.7.2
astunparse==1.6.3
attrs==23.1.0
Babel>=2.9.1
beautifulsoup4==4.12.2
cached-property==1.5.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-man==0.4.1
colorama==0.4.6
cryptography==3.4.8
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.8
Django==3.2.23
django-cors-headers==3.5.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
docopt==0.6.2
docutils==0.16
drf-yasg==1.21.7
ebaysdk==2.1.5
filelock==3.12.2
Flask==2.2.5
flatbuffers==23.5.26
freezegun==0.3.11
future==0.18.3
gast==0.4.0
gevent==1.5.0
ghp-import==2.1.0
google-auth==2.25.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.60.0
h5py==3.8.0
idna==2.10
importlib-metadata==6.7.0
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.11.0
libclang==16.0.6
libsass==0.20.1
livereload==2.6.3
lunr==0.5.8
lxml==4.6.5
Markdown==3.4.4
MarkupSafe==2.1.3
mergedeep==1.3.4
mkdocs==1.5.3
mkdocs-material==9.2.7
mkdocs-material-extensions==1.2
mpmath==1.3.0
nltk==3.8.1
num2words==0.5.9
numpy==1.21.6
oauthlib==3.2.2
ofxparse==0.19
opt-einsum==3.3.0
packaging==23.2
paginate==0.5.6
pandas==1.3.5
passlib==1.7.4
pathspec==0.11.2
Pillow==9.0.1
platformdirs==4.0.0
polib==1.1.0
polls==0.1
protobuf==3.19.6
psutil==5.8.0
psycopg2==2.7.7
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydot==1.4.2
Pygments==2.16.1
pymdown-extensions==10.2.1
pyOpenSSL==20.0.1
pyparsing==3.1.1
PyPDF2==1.26.0
pyserial==3.5
python-dateutil==2.8.1
python-ldap==3.4.0
python-stdnum==1.16
pytz==2023.3.post1
pyusb==1.0.2
PyYAML==6.0.1
pyyaml_env_tag==0.1
qrcode==6.1
regex==2022.10.31
reportlab==3.5.59
requests>=2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
sympy==1.10.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.66.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.5
virtualenv==20.25.0
vobject==0.9.6.1
watchdog==3.0.0
Werkzeug==2.2.3
wrapt==1.16.0
xlrd==1.1.0
XlsxWriter==1.1.2
xlwt==1.3.0
zeep==4.0.0
zipp==3.15.0